+++ src/project.cpp
@@ -2410,7 +2410,7 @@ QString Project::fixGraphicPath(QString path) {
     return path;
 }
 
-QString Project::getScriptFileExtension(bool usePoryScript) {
+QString Project::getScriptFileExtension(bool usePoryScript) const {
     if(usePoryScript) {
         return ".pory";
     } else {
@@ -2418,11 +2418,48 @@ QString Project::getScriptFileExtension(bool usePoryScript) {
     }
 }
 
-QString Project::getScriptDefaultString(bool usePoryScript, QString mapName) {
+QString Project::getScriptDefaultString(bool usePoryScript, QString mapName) const {
     if(usePoryScript)
         return QString("mapscripts %1_MapScripts {}").arg(mapName);
     else
         return QString("%1_MapScripts::\n\t.byte 0\n").arg(mapName);
+}
+
+QString Project::getMapScriptsFilePath(const QString &mapName) const {
+    const bool usePoryscript = projectConfig.getUsePoryScript();
+    auto path = QDir::cleanPath(root + "/data/maps/" + mapName + "/scripts");
+    auto extension = getScriptFileExtension(usePoryscript);
+    if (usePoryscript && !QFile::exists(path + extension))
+        extension = getScriptFileExtension(false);
+    path += extension;
+    return path;
+}
+
+QStringList Project::getEventScriptsFilePaths() const {
+    QStringList filePaths(QDir::cleanPath(root + "/data/event_scripts.s"));
+    const QString scriptsDir = QDir::cleanPath(root + "/data/scripts");
+    const QString mapsDir = QDir::cleanPath(root + "/data/maps");
+    const bool usePoryscript = projectConfig.getUsePoryScript();
+
+    if (usePoryscript) {
+        QDirIterator it_pory_shared(scriptsDir, {"*.pory"}, QDir::Files);
+        while (it_pory_shared.hasNext())
+            filePaths << it_pory_shared.next();
+
+        QDirIterator it_pory_maps(mapsDir, {"scripts.pory"}, QDir::Files, QDirIterator::Subdirectories);
+        while (it_pory_maps.hasNext())
+            filePaths << it_pory_maps.next();
+    }
+
+    QDirIterator it_inc_shared(scriptsDir, {"*.inc"}, QDir::Files);
+    while (it_inc_shared.hasNext())
+        filePaths << it_inc_shared.next();
+
+    QDirIterator it_inc_maps(mapsDir, {"scripts.inc"}, QDir::Files, QDirIterator::Subdirectories);
+    while (it_inc_maps.hasNext())
+        filePaths << it_inc_maps.next();
+
+    return filePaths;
 }
 
 void Project::loadEventPixmaps(QList<Event*> objects) {