+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The **"Breaking Changes"** listed below are changes that have been made in the d
 - Add ruler tool for measuring metatile distance in events tab (Right-click to turn on/off, left-click to lock in place).
 - Add delete button to wild pokemon encounters tab.
 - Add shortcut customization via `Options -> Edit Shortcuts`.
+- Add custom text editor commands in `Options -> Edit Preferences`, a tool-button next to the `Script` combo-box, and `Tools -> Open Project in Text Editor`. The tool-button will open the containing file to the cooresponding script.
 
 ### Changed
 - Holding `shift` now toggles "Smart Path" drawing; when the "Smart Paths" checkbox is checked, holding `shift` will temporarily disable it.